+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 2023-06-06
-1. FEAT: `pontx-config.json`   new parameters `plugins.fetch.options.translate.translateCacheDir`，Caching translation results
+1. FEAT: `pontx-config.json`   new parameters `plugins.fetch.options.translate.cacheFilePath`，Caching translation results
 
 ## 2023-06-03
 1. FIX: update baidu translate lib